Reorganise static files
Everything now lives in `/static`:
- `to_compile`: files that will be compiled by Parcel before being collected
- `to_collect`: files that will be directly collected
- `compiled`: files compiled by Parcel, to be collected
- `collected`: all collected files

Compiles the two last JS files that were not compiled by Parcel
