+++ .prettierignore
@@ -5,8 +5,8 @@
 __pycache__
 .venv
 .parcel-cache
-staticfiles
-compiled_static
+static/collected
+static/compiled
 node_modules
 
 # Directories that include a .gitignore
