+++ static/to_collect/img/publication_tree_example.png
