+++ static/to_compile/entrypoints/admin-map-widget.js
