+++ static/to_compile/entrypoints/publication.js
