+++ static/to_compile/src/map_controller.js
