+++ templates/publication_upload.html
@@ -53,6 +53,6 @@
         Générer la publication</button>
 </form>
 
-<script src="{% static 'js/publication.js' %}" defer></script>
+<script src="{% static 'publication.js' %}" defer></script>
 
 {% endblock %}
